Moved direction finding methods into world.
The direction finding functions are now supported by world, accepting either arbitrary locations, or ants as inputs, using a decorator to convert ants to locations. The old method of asking an ant for these things is still supported by calling the world methods, so old code should hopefully not break. Also added a closest_friend method that finds the closest friend to a given location that is not in that location.